Making Progress
So, Graphing is kind of messed up, but that's okay because the real graphing will be done elsewhere, with less data.
Need to get OHLC data of swings to the ElliotAnalyzer, or to the handler. Probably handler.
Finally got wave2() working.
Next up is fixing graphing issues. and annotating graph such that wave identifiers pop up